requirements.txt
flask==3.1.3
python-dotenv==1.2.2
psycopg2-binary==2.9.11
werkzeug==3.1.8
